+++ CHANGELOG.md
@@ -13,3 +13,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 
 - Syntax changes
+- Polish attributes before sending them